+++ libp2p/abc.py
@@ -1,5 +1,3 @@
-# host_interface
-
 from abc import (
     ABC,
     abstractmethod,
@@ -8,7 +6,6 @@ from collections.abc import (
     Sequence,
 )
 from typing import (
-    TYPE_CHECKING,
     Any,
     AsyncContextManager,
 )
@@ -47,9 +44,6 @@ from libp2p.tools.async_service import (
 from libp2p.transport.listener_interface import (
     IListener,
 )
-
-if TYPE_CHECKING:
-    from libp2p.network.notifee_interface import INotifee  # noqa: F401
 
 # network_interface
 
@@ -112,6 +106,53 @@ class INetwork(ABC):
 
 class INetworkService(INetwork, ServiceAPI):
     pass
+
+
+# notifee_interface
+
+
+class INotifee(ABC):
+    @abstractmethod
+    async def opened_stream(self, network: "INetwork", stream: INetStream) -> None:
+        """
+        :param network: network the stream was opened on
+        :param stream: stream that was opened
+        """
+
+    @abstractmethod
+    async def closed_stream(self, network: "INetwork", stream: INetStream) -> None:
+        """
+        :param network: network the stream was closed on
+        :param stream: stream that was closed
+        """
+
+    @abstractmethod
+    async def connected(self, network: "INetwork", conn: INetConn) -> None:
+        """
+        :param network: network the connection was opened on
+        :param conn: connection that was opened
+        """
+
+    @abstractmethod
+    async def disconnected(self, network: "INetwork", conn: INetConn) -> None:
+        """
+        :param network: network the connection was closed on
+        :param conn: connection that was closed
+        """
+
+    @abstractmethod
+    async def listen(self, network: "INetwork", multiaddr: Multiaddr) -> None:
+        """
+        :param network: network the listener is listening on
+        :param multiaddr: multiaddress listener is listening on
+        """
+
+    @abstractmethod
+    async def listen_close(self, network: "INetwork", multiaddr: Multiaddr) -> None:
+        """
+        :param network: network the connection was opened on
+        :param multiaddr: multiaddress listener is no longer listening on
+        """
 
 
 # host_interface
