+++ libp2p/pubsub/pubsub_notifee.py
@@ -9,12 +9,10 @@ import trio
 
 from libp2p.abc import (
     INetwork,
+    INotifee,
 )
 from libp2p.network.connection.net_connection_interface import (
     INetConn,
-)
-from libp2p.network.notifee_interface import (
-    INotifee,
 )
 from libp2p.network.stream.net_stream_interface import (
     INetStream,
